+++ tests/middleware/test_exception_handler_middleware.py
@@ -97,7 +97,9 @@ def test_exception_handler_middleware_exception_handlers_mapping() -> None:
         return Response(content={"an": "error"}, status_code=HTTP_500_INTERNAL_SERVER_ERROR, media_type=MediaType.JSON)
 
     app = Starlite(route_handlers=[handler], exception_handlers={Exception: exception_handler}, openapi_config=None)
-    assert app.route_map["/"]["_asgi_handlers"]["GET"]["asgi_app"].exception_handlers == {Exception: exception_handler}
+    assert app.asgi_router.route_map["/"]["_asgi_handlers"]["GET"]["asgi_app"].exception_handlers == {
+        Exception: exception_handler
+    }
 
 
 def test_exception_handler_middleware_calls_app_level_after_exception_hook() -> None:
